chore(shared): move rich preview assets
