+++ projects/@shared/components/rich-previews/js/utils.js
@@ -1,0 +1,8 @@
+export const getHostname = url => {
+  if(!url) return '';
+  const hostname = (new URL(url).hostname);
+  const wwwPrefix = 'www.';
+  return hostname.startsWith(wwwPrefix)
+    ? hostname.slice(wwwPrefix.length)
+    : hostname;
+};
