+++ projects/@shared/components/rich-previews/linkedin-preview.vue
@@ -51,7 +51,7 @@
 
 <script>
 import { computed } from 'vue';
-import { getHostname } from '../../../rich-previews/@shared/js/utils.js';
+import { getHostname } from './js/utils.js';
 
 export default {
   props: {
@@ -84,7 +84,7 @@ export default {
 </script>
 
 <style>
-@import "../../../rich-previews/@shared/css/defaults.css";
+@import "./css/defaults.css";
 
 :root {
   --linkedin-color-border: #E1E8ED;
@@ -169,7 +169,7 @@ export default {
 
 .linkedin-preview__image-fallback {
   height: 100%;
-  background-image: url('../../../rich-previews/src/@assets/image-not-found.svg');
+  background-image: url('/assets/image-not-found.svg');
   background-repeat: no-repeat;
   background-position: center;
   background-size: 3em;
